hot fix in alembic-generated tables
