--- alembic/versions/27f10b50ffe0_create_posts_table.py
@@ -1,25 +1,0 @@
-"""create posts table
-
-Revision ID: 27f10b50ffe0
-Revises: 
-Create Date: 2022-02-04 18:11:03.144546
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '27f10b50ffe0'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    op.create_table('posts', sa.Column('id', sa.Integer(), nullable=False,
-                    primary_key=True), sa.Column('title', sa.String(), nullable=False))
-
-
-def downgrade():
-    op.drop_table('posts')
