--- alembic/versions/71c8e245a631_auto_update_user_table.py
@@ -1,28 +1,0 @@
-"""auto update user table
-
-Revision ID: 71c8e245a631
-Revises: b7587b88fe32
-Create Date: 2022-02-24 21:48:24.921607
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '71c8e245a631'
-down_revision = 'b7587b88fe32'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('phone_number', sa.String(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'phone_number')
-    # ### end Alembic commands ###
