--- alembic/versions/b7587b88fe32_auto_vote_table.py
@@ -1,34 +1,0 @@
-"""auto vote table
-
-Revision ID: b7587b88fe32
-Revises: 75197dbb07dc
-Create Date: 2022-02-24 13:29:11.886479
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'b7587b88fe32'
-down_revision = '75197dbb07dc'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('votes',
-    sa.Column('post_id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['post_id'], ['posts.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('post_id', 'user_id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('votes')
-    # ### end Alembic commands ###
