--- alembic/versions/c8557cebb225_add_fk_to_posts_table.py
@@ -1,27 +1,0 @@
-"""add FK to posts table
-
-Revision ID: c8557cebb225
-Revises: 9f3c0ce4ec3d
-Create Date: 2022-02-24 12:57:56.711505
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'c8557cebb225'
-down_revision = '9f3c0ce4ec3d'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    op.add_column('posts', sa.Column('owner_id', sa.Integer(), nullable=False))
-    op.create_foreign_key('posts_users_fk', source_table='posts', referent_table='users',
-                          local_cols=['owner_id'], remote_cols=['id'], ondelete='CASCADE')
-
-
-def downgrade():
-    op.drop_constraint('posts_users_fk', table_name='posts')
-    op.drop_column('posts', 'owner_id')
